requirements.txt
scrapy
readability-lxml
markdownify
lxml[html_clean]